+++ NEWS.md
@@ -1,5 +1,7 @@
 # 0.6.0.9000
 
+* New function `getPlottingData()` to return the data required by the first
+argument of the custom plotting functions.
 
 # 0.6.0